fix: add ability to specify subject for report
This commit adds the ability to specify a subject for the report in the `start_report` function. If a subject is provided as an argument, it will be used directly without prompting the user. Otherwise, the user will be prompted to enter a subject.

This change addresses the issue of not being able to provide a subject for the report, which was reported in issue #36.

Fixes #36